+++ src/services/payment/shipping.js
@@ -1,7 +1,14 @@
-function amount() {
-  return parseInt(parseFloat(process.env.SHIPPING_AMOUNT).toFixed(2) * 100)
+function amount(total) {
+  return !isFreeShipping(total)
+    ? parseInt(parseFloat(process.env.SHIPPING_AMOUNT).toFixed(2) * 100)
+    : 0
+}
+
+function isFreeShipping(total) {
+  return total >= process.env.FREE_SHIPPING_SUBTOTAL
 }
 
 module.exports = {
-  amount
+  amount,
+  isFreeShipping
 }